+++ src/Model.hs
@@ -17,6 +17,7 @@ import Vehicles.Components
 import People.Data
 import Research.Data
 import Resources ( RawResource(..), Biological(..), Chemical(..), Mechanical(..) )
+import Space.Data
 
 -- You can define all of your database entities in the entities file.
 -- You can find more information on persistent and how to declare entities